+++ archive_download.py
@@ -1,4 +1,5 @@
 import os
+import py7zr
 import PySimpleGUI as sg
 import common_functions as common
 from urllib.parse import quote
@@ -9,7 +10,7 @@ from http.client import HTTPResponse
 _base_url = "https://archive.org"
 
 
-def download_files(platform_id: str, filenames: list, output_folder: str, cb_progressbar: sg.ProgressBar, cb_statusbar: sg.StatusBar, cb_text_status: sg.Text, cb_function):
+def download_files(platform_id: str, filenames: list, output_folder: str, unzip: bool, cb_progressbar: sg.ProgressBar, cb_statusbar: sg.StatusBar, cb_text_status: sg.Text, cb_function):
   def download(filename: str):
     filename_quoted = quote(filename)
     full_url = f"{_base_url}/download/{platform_id}/{filename_quoted}"
@@ -36,8 +37,16 @@ def download_files(platform_id: str, filenames: list, output_folder: str, cb_pro
   max = len(filenames)
   for filename in filenames:
     count+=1
+    str_counter = f" ({count}/{max})"
+    tronqued_filename = f"{filename[:30]}{'...' if len(filename) > 30 else ''}"
     cb_progressbar.update(0, 0)
     cb_text_status.update("0Kb / 0Kb")
-    cb_statusbar.update(f"Downloading... ({filename[:20]}{'...' if len(filename) > 20 else ''}) ({count}/{max})")
+    cb_statusbar.update(f"Downloading {tronqued_filename}" + str_counter)
     download(filename)
+    if unzip:
+      cb_statusbar.update(f"Extracting {tronqued_filename}" + str_counter)
+      py7zr.SevenZipFile(os.path.join(output_folder, filename)).extractall(output_folder)
+      cb_statusbar.update(f"Extraction completed! Deleting archive {tronqued_filename}" + str_counter)
+      os.remove(os.path.join(output_folder, filename))
+  
   cb_function(max)